+++ packages/blade-svelte/src/components/index.ts
@@ -25,6 +25,11 @@ export { default as Amount } from './Amount/Amount.svelte';
 
 // Icons
 export * from './Icons';
+// Badge
+export { default as Badge } from './Badge/Badge.svelte';
+
+// Divider
+export { default as Divider } from './Divider/Divider.svelte';
 
 // BladeProvider
 export * from './BladeProvider';