+++ packages/blade/src/components/Charts/BarChart/types.ts
@@ -36,6 +36,15 @@ type ChartBarProps = {
    */
   showLegend?: boolean;
   /**
+   * Whether to hide the bar.
+   *
+   */
+  hide?: boolean;
+  /**
+   * The width of the bar in pixels.
+   */
+  barSize?: RechartsBarProps['barSize'];
+  /**
    * The index of the bar chart.
    * @private
    */